backend/identifyFlask/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
Pillow==10.1.0
scipy==1.11.3
tetra3 @ git+https://github.com/esa/tetra3.git@f98bb4f769f6054cdc1fbc023a07934700ceb70e
Werkzeug==3.0.1
zipp==3.17.0
